implement parsing of rule bodies
this needed some adaptions in the scanner to account for the fact that we
advance it in two places (the Parse() function and the parseRuleOrVariable()
function). So we can't just rely on always advancing the scanner in the
Parse() loop anymore. Since the bufio.Scanner doesn't really provide a way to
check if scanning is finished without calling Scan() the MakefileScanner
implements an attribute Finished that can be checked.

The parser addition is basically just another loop in the case where we found
a rule that keeps adding to the rule body array as long as the line matches
the rule body regex, advancing the scanner on the way.

This commit also adds whitespace trimming to all parsed attributes.