add rule to make sure timestamp variables are simply expanded
this rule checks for variables that potentially contain a timestamp and makes
sure they are simply expanded thus staying the same between rule executions.
The way it determines whether it might be a timestamp variable is pretty naive
right now as it just checks for the substring 'date' in the variable
assignment.

fixes #17