make checkmake go install-able
this makes it so the project can be installed simply via `go install` to
conform with golang standards. The base checkmake.go is just a symlink
to the actual main binary in `cmd/checkmake` so we don't need to
maintain the same file in two locations.

There is one downside with using go install right now, which is that all
of the version information for `checkmake --version` will be missing as
it gets filled in by make on build. We might find a way to fix this in
the future but for now `make` or the prebuilt packages are the preferred
way of installing checkmake. And `go install` is the less nice version
to conform with standards and support easier usage from other tools.

fixes #19